refactor: move AgentDelegation plugin to agent directory
